dashdec: Fix seg num decrease in refresh_manfiest
It is possible for the demuxer to be processing at the live edge and waiting for
a segment in the future.  When that happens 'cur_seq_no' can be past the end of
the timelines indicated by the MPD.

The functions used in refresh_manifest() that attempt to calculate the next
segment number based on the timeline data will end up clamping to the end of
the MPD, which would cause the segment number to go backwards.

To prevent this, we make sure the segment number never decreases when
we're past the documented segments in the MPD.