dashdec: representation filtering and live edge time reporting (#50)
* dashdec: representation filter, add liveEdgeSegStartTime

Skip parsing representations for audio/video streams that we're not processing. This avoids any additional overhead housekeeping that is done for streams referenced by the MPD. This was a patch we had applied years ago that got lost in ffmpeg updates.

Add the start time of the segment at the live edge in the packet metadata.